+++ apps/web/src/components/DevtoolsArrow.tsx
@@ -1,7 +1,7 @@
 import { AnimatePresence, motion } from "framer-motion";
-import { TrackingEvent } from "lib/tracking";
+import { useTracking } from "lib/tracking";
 import { CornerRightDown } from "lucide-react";
-import { usePlausible } from "next-plausible";
+
 import { useState, useEffect } from "react";
 
 const DEVTOOLS_ID = "abby-devtools-collapsed";
@@ -10,6 +10,7 @@ export function useDevtoolsPosition() {
   const [devtoolsPosition, setDevtoolsPosition] = useState<DOMRect | null>(
     null
   );
+  const trackEvent = useTracking();
 
   useEffect(() => {
     const devtools = document.getElementById(DEVTOOLS_ID);
@@ -28,10 +29,16 @@ export function useDevtoolsPosition() {
 
     setDevtoolsPosition(devtools.getBoundingClientRect());
 
+    const devtoolsAnalytics = () => {
+      trackEvent("Devtools Opened");
+    };
+    devtools.addEventListener("click", devtoolsAnalytics);
+
     return () => {
       resizeObserver.disconnect();
+      devtools.removeEventListener("click", devtoolsAnalytics);
     };
-  }, []);
+  }, [trackEvent]);
 
   // listen to window resize
   useEffect(() => {
@@ -54,7 +61,7 @@ export function useDevtoolsPosition() {
 }
 
 export function DevtoolsArrow() {
-  const plausible = usePlausible();
+  const trackEvent = useTracking();
   const devtoolsPosition = useDevtoolsPosition();
 
   useEffect(() => {
@@ -64,17 +71,17 @@ export function DevtoolsArrow() {
 
       switch (messageType) {
         case "abby:update-flag": {
-          plausible(TrackingEvent.DEVTOOLS_INTERACTION, {
+          trackEvent("Devtools Interaction", {
             props: {
-              type: "update-flag",
+              type: "Flag Updated",
             },
           });
           break;
         }
         case "abby:select-variant": {
-          plausible(TrackingEvent.DEVTOOLS_INTERACTION, {
+          trackEvent("Devtools Interaction", {
             props: {
-              type: "select-variant",
+              type: "Variant Selected",
             },
           });
           break;
@@ -87,7 +94,7 @@ export function DevtoolsArrow() {
     return () => {
       window.removeEventListener("message", onMessage);
     };
-  }, []);
+  }, [trackEvent]);
 
   return (
     <AnimatePresence>