+++ apps/web/src/components/Pricing.tsx
@@ -2,13 +2,14 @@ import clsx from "clsx";
 import { cn } from "lib/utils";
 import { Info } from "lucide-react";
 import { useSession } from "next-auth/react";
+
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { AiOutlineCheckCircle } from "react-icons/ai";
 import { getLimitByPlan } from "server/common/plans";
+import { Plan } from "types/plausible-events";
 import { Tooltip, TooltipContent, TooltipTrigger } from "./Tooltip";
-import { usePlausible } from "next-plausible";
-import { TrackingEvent } from "lib/tracking";
+import { useTracking } from "lib/tracking";
 
 type PricingElementProps = {
   price: string;
@@ -20,6 +21,7 @@ type PricingElementProps = {
   ctaText?: string;
   href: string;
   isFull?: boolean;
+  planName: Plan;
 };
 
 function PricingElement({
@@ -31,8 +33,10 @@ function PricingElement({
   ctaText,
   href,
   isFull,
+  planName,
 }: PricingElementProps) {
-  const plausible = usePlausible();
+  const trackEvent = useTracking();
+
   return (
     <div
       className={cn(
@@ -58,18 +62,14 @@ function PricingElement({
       >
         <Link
           href={href}
-          onClick={() =>
-            plausible(TrackingEvent.PRICING_CLICKED, {
-              props: {
-                title,
-              },
-            })
-          }
           className={cn(
             "my-6 w-full rounded-xl border px-4 py-2 transition-colors duration-200 ease-in-out",
             "border-accent-background hover:bg-accent-background",
             isFull && "lg:order-2 lg:w-64 lg:justify-self-center"
           )}
+          onClick={() =>
+            trackEvent("Plan Selected", { props: { Plan: planName } })
+          }
         >
           {ctaText ?? `Choose ${title}`}
         </Link>
@@ -123,6 +123,7 @@ export function PricingTable() {
           href={session.status === "authenticated" ? "/projects" : "/login"}
           price="Free"
           title="Hobby"
+          planName="HOBBY"
           subtitle="Good for IndieHackers that want to get started with A/B Testing & Feature Flags. No Credit card required"
           features={[
             `${basePlan.eventsPerMonth.toLocaleString()} Events / month`,
@@ -136,6 +137,7 @@ export function PricingTable() {
           price="12€"
           title="Startup"
           subtitle="Optimal for startups & small businesses that want to dive deeper with A/B Testing & Feature Flags"
+          planName="STARTUP"
           features={[
             `${startupPlan.eventsPerMonth.toLocaleString()} Events / month`,
             `${startupPlan.tests} A/B Tests`,
@@ -149,6 +151,7 @@ export function PricingTable() {
           href={session.status === "authenticated" ? "/projects" : "/login"}
           price="89€"
           title="Pro"
+          planName="PRO"
           subtitle="Perfect for growing companies that want to scale their A/B Testing & Feature Flags and get more insights"
           features={[
             `${proPlan.eventsPerMonth.toLocaleString()} Events / month`,
@@ -162,6 +165,7 @@ export function PricingTable() {
           href={"/contact"}
           price="Talk to us!"
           title="Enterprise"
+          planName="ENTERPRISE"
           subtitle="For even the biggest enterprise companies."
           ctaText="Contact us"
           features={[