+++ apps/web/src/lib/tracking.ts
@@ -1,6 +1,4 @@
-export const enum TrackingEvent {
-  DEVTOOLS_OPENED = "devtools-opened",
-  PRICING_CLICKED = "pricing-clicked",
-  DEVTOOLS_INTERACTION = "devtools-interaction",
-  AB_TESTING_GENERATOR = "ab-testing-generator",
-}
+import { usePlausible as _usePlausible } from "next-plausible";
+import { PlausibleEvents } from "types/plausible-events";
+
+export const useTracking = _usePlausible<PlausibleEvents>;