+++ apps/web/src/pages/index.tsx
@@ -20,11 +20,14 @@ import { generateCodeSnippets } from "utils/snippets";
 import abbyScreenshot from "../../public/screenshot.png";
 import { NextPageWithLayout } from "./_app";
 import { DevtoolsArrow } from "components/DevtoolsArrow";
+import { PlausibleEvents } from "types/plausible-events";
+import { useTracking } from "lib/tracking";
 
 const { useAbby, AbbyProvider, useFeatureFlag, __abby__, withDevtools } =
   createAbby({
     projectId: "clk8ld04v0000l0085dqsxpsr",
     currentEnvironment: "production",
+    environments: ["production", "development"],
     tests: {
       SignupButton: {
         variants: ["A", "B", "C"],
@@ -64,6 +67,8 @@ const Home: NextPageWithLayout<
   const { onAct, variant } = useAbby("SignupButton");
 
   const forceDarkTheme = useFeatureFlag("ForceDarkTheme");
+
+  const trackEvent = useTracking();
 
   useEffect(() => {
     if (forceDarkTheme) {
@@ -106,7 +111,10 @@ const Home: NextPageWithLayout<
           <div className="flex flex-col items-center">
             <Link
               href="/login"
-              onClick={() => onAct()}
+              onClick={() => {
+                trackEvent("Sign Up Clicked");
+                onAct();
+              }}
               className={twMerge(
                 "mt-12 rounded-xl bg-accent-background px-6 py-2 text-xl font-semibold text-accent-foreground no-underline transition-transform duration-150 ease-in-out hover:scale-110"
               )}