+++ packages/svelte/src/tests/useAbby.test.ts
@@ -1,23 +1,17 @@
 import { TestStorageService } from "../lib/StorageService";
-import { HttpService, AbbyEventType } from "@tryabby/core";
+import { HttpService, AbbyEventType, getABStorageKey } from "@tryabby/core";
 import { createAbby } from "../lib/createAbby";
-import {
-  it,
-  describe,
-  expect,
-  afterEach,
-  vi,
-  beforeAll,
-  afterAll,
-} from "vitest";
+import { it, describe, expect, afterEach, vi, beforeAll, afterAll } from "vitest";
 /// @ts-ignore it doesn't have types
 import { get } from "svelte/store";
+import Cookies from "js-cookie";
 
 const OLD_ENV = process.env;
 
 describe("useAbby working", () => {
   it("returns a valid variant", () => {
     const { useAbby } = createAbby({
+      environments: [],
       projectId: "123",
       tests: {
         test: { variants: ["OldFooter", "NewFooter"] },
@@ -40,6 +34,7 @@ describe("useAbby working", () => {
 
     getSpy.mockReturnValue(persistedValue);
     const { useAbby } = createAbby({
+      environments: [],
       projectId: "123",
       tests: {
         test: { variants },
@@ -56,9 +51,35 @@ describe("useAbby working", () => {
     expect(result).toEqual(persistedValue);
   });
 
+  it("should look up the return value", () => {
+    const persistedValue = "SimonsText";
+    const variants = ["SimonsText", "MatthiasText", "TomsText", "TimsText"] as const;
+
+    const getSpy = vi.spyOn(TestStorageService, "get");
+
+    getSpy.mockReturnValue(persistedValue);
+    const { useAbby } = createAbby({
+      environments: [],
+      projectId: "123",
+      tests: {
+        test: { variants },
+      },
+    });
+
+    const { variant } = useAbby("test", {
+      SimonsText: "a",
+      MatthiasText: "b",
+      TomsText: "c",
+      TimsText: "d",
+    });
+    const result = get(variant);
+    expect(result).toBe("a");
+  });
+
   it("should ping the current info on mount", () => {
     const spy = vi.spyOn(HttpService, "sendData");
     const { useAbby } = createAbby({
+      environments: [],
       projectId: "123",
       tests: {
         test: { variants: ["A", "B", "C"] },
@@ -68,14 +89,13 @@ describe("useAbby working", () => {
     const { variant, onAct } = useAbby("test");
 
     //onAct();
-    expect(spy).toHaveBeenCalledWith(
-      expect.objectContaining({ type: AbbyEventType.PING })
-    );
+    expect(spy).toHaveBeenCalledWith(expect.objectContaining({ type: AbbyEventType.PING }));
   });
 
   it("should notify the server with onAct", () => {
     const spy = vi.spyOn(HttpService, "sendData");
     const { useAbby } = createAbby({
+      environments: [],
       projectId: "123",
       tests: {
         test: { variants: ["A", "B", "C"] },
@@ -85,12 +105,10 @@ describe("useAbby working", () => {
     const { variant, onAct } = useAbby("test");
 
     onAct();
-    expect(spy).toHaveBeenCalledWith(
-      expect.objectContaining({ type: AbbyEventType.PING })
-    );
+    expect(spy).toHaveBeenCalledWith(expect.objectContaining({ type: AbbyEventType.PING }));
   });
 
-  it("returns the correct possible variant values", () => {
+  it("returns the correct possible variant values", async () => {
     const expectedVariants: readonly string[] = [
       "SimonsText",
       "MatthiasText",
@@ -98,6 +116,7 @@ describe("useAbby working", () => {
       "TimsText",
     ];
     const { getVariants } = createAbby({
+      environments: [],
       projectId: "123",
       tests: {
         test: {
@@ -107,5 +126,28 @@ describe("useAbby working", () => {
     });
     const recievedVariants = get(getVariants("test"));
     expect(recievedVariants).toEqual(expectedVariants);
+  });
+
+  it("uses lookup object when retrieving variant", async () => {
+    const { getABTestValue } = createAbby({
+      environments: [],
+      projectId: "123",
+      tests: {
+        lookupTest: {
+          variants: ["SimonsText", "MatthiasText", "TomsText", "TimsText"],
+        },
+      },
+    });
+
+    const pickedVariant = getABTestValue("lookupTest");
+    const lookupMap = {
+      SimonsText: 1,
+      MatthiasText: 2,
+      TomsText: 3,
+      TimsText: 4,
+    };
+
+    const value = getABTestValue("lookupTest", lookupMap);
+    expect(value).toBe(lookupMap[pickedVariant]);
   });
 });