+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @tryabby/angular
 
+## 1.2.0
+
+### Minor Changes
+
+- add lookup object to useAbby and ab test value function
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/core@4.2.0
+
+## 1.1.0
+
+### Minor Changes
+
+- add defineConfig helper to ease the usage with the cli
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/core@4.1.0
+
 ## 1.0.1
 
 ### Patch Changes