+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tryabby/core
 
+## 4.2.0
+
+### Minor Changes
+
+- add lookup object to useAbby and ab test value function
+
+## 4.1.0
+
+### Minor Changes
+
+- add defineConfig helper to ease the usage with the cli
+
 ## 4.0.1
 
 ### Patch Changes